refactor(dev): separeate composeUp/Down from showDb
